apps/api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.12

# DB
sqlalchemy==2.0.36
alembic==1.14.0
psycopg[binary]==3.2.3

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
email-validator==2.2.0

# Settings
pydantic-settings==2.6.1
